flask-api-for-chatbot/flask-api-for-chatbot/requirements.txt
streamlit 
flask 
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai